TASK-52 [Rename]: 내 대시보드 & 초대 응답 초기 값 파일 분리
