+++ src/lib/constants/invitationsInit.ts
@@ -1,13 +1,3 @@
-export const INIT_MYDASHBOARDS_CONTEXT = {
-  fetchedDashboards: [],
-  localDashboards: [],
-  loading: true,
-  error: null,
-  addDashboard: () => {},
-  fetchDashboards: async () => {},
-};
-
-// 초대받은 대시보드 목록
 export const INIT_INVITATIONS_REQUEST = {
   size: 10,
 };
@@ -37,17 +27,4 @@ export const INIT_INVITATIONS_RESPONSE = {
       updatedAt: new Date().toISOString(),
     },
   ],
-};
-
-// 나의 대시보드 목록
-export const INIT_DASHBOARDS_REQUEST = {
-  navigationMethod: 'pagination',
-  page: 1,
-  size: 10,
-};
-
-export const INIT_DASHBOARDS_RESPONSE = {
-  dashboards: [],
-  totalCount: 0,
-  cursorId: null,
 };
